requirements.txt
# Core OCR and document processing
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# Data analysis and graph processing
networkx>=3.1
pandas>=2.0.0

# Natural Language Processing
spacy>=3.7.0

# Environment configuration
python-dotenv>=1.0.0

# Database
# SQLite is included in Python standard library

# Additional useful dependencies for document processing
pypdf2>=3.0.0
